requirements.txt
colorama==0.4.3
Django==2.1.5
django-cors-headers==3.4.0
labelImg==1.8.3
lxml==4.5.1
numpy==1.19.0
opencv-python-headless==4.2.0.34
Pillow==7.1.2
PyQt5==5.15.0
PyQt5-sip==12.8.0
python-decouple==3.3
pytz==2020.1
qrcode==6.1
six==1.15.0
tensorflow==1.14.0
keras==2.2.4